Experiment with rendering an FPM
@logankilpatrick take a look at this, there's a list of todos at the 
bottom.  Essentially, all we need to do is to obtain some data source 
for those, and then do the Markdown plots and stuff (but those can be 
done separately).

I've taken a dependency on Mustache, which is a Julia template engine.  
If we can use that, it'd be really easy to template and generate a 
report.